refactor(Observable): remove Observable
using a subscription function makes observable slow
