--- chore/rfc.js
@@ -1,10 +1,0 @@
-/**
- * Created by tushar.mathur on 05/10/16.
- */
-
-'use strict'
-
-const ObservableTests = require('es-observable-tests')
-const {Observable} = require('../src/Observable')
-
-ObservableTests.runTests(Observable)
