+++ package.json
@@ -20,7 +20,6 @@
     "benchmark": "^2.1.1",
     "coveralls": "^2.11.14",
     "cz-conventional-changelog": "^1.2.0",
-    "es-observable-tests": "^0.3.0",
     "ghooks": "^1.3.2",
     "nyc": "^8.3.1",
     "semantic-release": "^4.3.5",
